move get_default_task_file, in a common utils_executors
